LayoutTests/imported/w3c/web-platform-tests/preload/preload-csp.sub-expected.txt

CONSOLE MESSAGE: Refused to load localhost:8801/preload/resources/dummy.js because it does not appear in the style-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/dummy.css because it does not appear in the style-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/square.png because it does not appear in the img-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/fonts/CanvasTest.ttf because it does not appear in the font-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/white.mp4 because it does not appear in the media-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/sound_5.oga because it does not appear in the media-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/foo.vtt because it does not appear in the media-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load localhost:8801/preload/resources/dummy.js because it does not appear in the style-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/dummy.css because it does not appear in the style-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/square.png because it does not appear in the img-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/fonts/CanvasTest.ttf because it does not appear in the font-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/white.mp4 because it does not appear in the media-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/sound_5.oga because it does not appear in the media-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/foo.vtt because it does not appear in the media-src directive of the Content Security Policy.
CONSOLE MESSAGE: <link rel=preload> must have a valid `as` value
CONSOLE MESSAGE: <link rel=preload> cannot have the empty string as `as` value

PASS Makes sure that preload requests respect CSP
